pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aurora-asr"
version = "1.0.0"
description = "Aurora ASR: OpenEnv-v4 Compliant Self-Healing CI/CD Agent"
authors = [
    {name = "Aurora Contributor"}
]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "gym>=0.21.0",
    "gymnasium>=0.27.0",
    "tree-sitter>=0.20.0",
    "pytest>=7.0.0",
    "numpy>=1.21.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "openenv-core>=0.2.0",
    "rich>=13.0.0"
]

[project.scripts]
server = "server.app:main"
